requirements.txt
streamlit==1.50.0
langchain==0.3.27
langchain-google-genai==2.1.9
langchain-core==0.3.72
python-dotenv==1.1.1